+++ packages/runtime/adapters/eventstore-adapters/eventstore-postgresql/src/reset-replication.ts
@@ -1,5 +1,5 @@
 import type { AdapterPool } from './types'
-import type { ReplicationStatus } from '@resolve-js/eventstore-base'
+import type { ReplicationState } from '@resolve-js/eventstore-base'
 
 const resetReplication = async (pool: AdapterPool): Promise<void> => {
   const {
@@ -17,22 +17,22 @@ const resetReplication = async (pool: AdapterPool): Promise<void> => {
   const replicationStateTableNameAsId = escapeId(replicationStateTableName)
   const threadsTableAsId = escapeId(`${eventsTableName}-threads`)
 
-  const notStarted: ReplicationStatus = 'notStarted'
+  const notStarted: ReplicationState['statusAndData']['status'] = 'notStarted'
 
-  const statements = [
-    `TRUNCATE ${databaseNameAsId}.${eventsTableNameAsId}`,
-    `TRUNCATE ${databaseNameAsId}.${secretsTableNameAsId}`,
-    `UPDATE ${databaseNameAsId}.${threadsTableAsId} SET "threadCounter" = 0`,
-    `UPDATE ${databaseNameAsId}.${replicationStateTableNameAsId} SET
+  await executeStatement(`
+  BEGIN WORK;
+  LOCK ${databaseNameAsId}.${replicationStateTableNameAsId} IN ACCESS EXCLUSIVE MODE;
+  TRUNCATE ${databaseNameAsId}.${eventsTableNameAsId};
+  TRUNCATE ${databaseNameAsId}.${secretsTableNameAsId};
+  UPDATE ${databaseNameAsId}.${threadsTableAsId} SET "threadCounter" = 0;
+  UPDATE ${databaseNameAsId}.${replicationStateTableNameAsId} SET
       "Status" = ${escape(notStarted)},
       "StatusData" = NULL,
       "Iterator" = NULL,
       "SuccessEvent" = NULL,
-      "LockExpirationTime" = 0`,
-  ]
-  for (const statement of statements) {
-    await executeStatement(statement)
-  }
+      "LockExpirationTime" = 0,
+      "LockId" = NULL;
+  COMMIT WORK;`)
 }
 
 export default resetReplication