+++ CHANGELOG.md
@@ -1,5 +1,52 @@
 # Changelog
 
+## [Upcoming](https://github.com/reimagined/resolve/tree/HEAD)
+
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.33.8...HEAD)
+
+**Documentation:**
+
+- Update ES Adapter API Reference [\#2181](https://github.com/reimagined/resolve/pull/2181)
+
+## [V0.33.8](https://github.com/reimagined/resolve/tree/V0.33.8) (2021-12-29)
+
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.33.7...V0.33.8)
+
+**Enhancements:**
+
+- More detailed command execution monitoring [\#2186](https://github.com/reimagined/resolve/pull/2186)
+- Use Promise.allSettled in postgres eventstore adapter dispose [\#2184](https://github.com/reimagined/resolve/pull/2184)
+- Export events via event loader. Improved export tests [\#2183](https://github.com/reimagined/resolve/pull/2183)
+
+**Other changes:**
+
+- Fix manual cloud test workflow [\#2132](https://github.com/reimagined/resolve/pull/2132)
+
+## [V0.33.7](https://github.com/reimagined/resolve/tree/V0.33.7) (2021-12-20)
+
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.33.6...V0.33.7)
+
+**Other changes:**
+
+- Add eventstore adapter runtimeInfo and setReconnectionMode. Add connection tests [\#2179](https://github.com/reimagined/resolve/pull/2179)
+- Add lambda runtime execution info logging [\#2177](https://github.com/reimagined/resolve/pull/2177)
+- Rework replicaiton to correctly handle locks [\#2174](https://github.com/reimagined/resolve/pull/2174)
+- Improve PLV8 warning message [\#2171](https://github.com/reimagined/resolve/pull/2171)
+- Migrate from 'request' package to 'node-fetch' in runtime [\#2169](https://github.com/reimagined/resolve/pull/2169)
+- Mini benchmarks initial [\#2134](https://github.com/reimagined/resolve/pull/2134)
+
+## [V0.33.5](https://github.com/reimagined/resolve/tree/V0.33.5) (2021-12-02)
+
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.33.4...V0.33.5)
+
+**Bug fixes:**
+
+- Commands scheduled in sagas are triggered way too late [\#2157](https://github.com/reimagined/resolve/issues/2157)
+
+**Documentation:**
+
+- Revisit client API docs [\#2145](https://github.com/reimagined/resolve/pull/2145)
+
 ## [V0.33.4](https://github.com/reimagined/resolve/tree/V0.33.4) (2021-11-26)
 
 [Full Changelog](https://github.com/reimagined/resolve/compare/V0.33.3...V0.33.4)